[MOV] Core: rename VRange.ts into VSelection.ts and RangeNode.ts into MarkerNode.ts
